Item10460: Revised fix for expanding %MACROS
Expanded the macros in the LOOKELSEWHERWEBS a bit too early and it
prevented the default from taking effect.

Also removed global variables, and changed the lazy init to be sure init
was called for every request.  Changes in the Preference Settings were
not being reliably applied.  Would impact a fcgi or mod_perl
environment.

Added unit test for ACRONYM features.

git-svn-id: http://svn.foswiki.org/trunk/FindElsewherePlugin@13013 0b4bb1d4-4e5a-0410-9cc4-b2b747904278